logout arreglado
