+++ src/components/Aside.jsx
@@ -2,12 +2,13 @@ import React, { useState } from "react";
 import { useApiLogin } from "../hooks/useApiLogin";
 import { useNavigate, useLocation } from "react-router-dom";
 import CambiarContraseña from "../components/CambiarContraseña";
+import Logout from "./logout";
 
 export default function Aside() {
   const { user, logout } = useApiLogin();
   const navigate = useNavigate();
   const location = useLocation();
-
+  console.log(user);
   const [openModal, setOpenModal] = useState(false);
 
   const menuItems = [
@@ -20,11 +21,6 @@ export default function Aside() {
 
   const handleNavigation = (path) => {
     navigate(path);
-  };
-
-  const handleLogout = () => {
-    logout();
-    navigate("/");
   };
 
   const isActive = (path) => {
@@ -61,9 +57,7 @@ export default function Aside() {
                 {user?.email || "admin@funval.test"}
               </p>
             </div>
-            <div className="mt-1 bg-blue-100 text-blue-700 px-3 py-1 rounded-full text-[10px] 2xl:text-xs font-medium">
-              Administrador
-            </div>
+            <div className="mt-1 bg-blue-100 text-blue-700 px-3 py-1 rounded-full text-[10px] 2xl:text-xs font-medium"></div>
             <button
               onClick={() => setOpenModal(true)}
               className="mt-3 px-4 py-2 bg-blue-600 text-white rounded-lg text-xs 2xl:text-sm font-medium hover:bg-blue-700 transition-colors shadow-md"
@@ -114,26 +108,7 @@ export default function Aside() {
             ))}
           </ul>
           <div className="flex justify-center">
-            <button
-              onClick={handleLogout}
-              className=" w-40 2xl:w-full bg-gradient-to-r from-red-500 to-red-600 text-white p-2 2xl:py-3 rounded-xl text-xs 2xl:text-sm font-medium hover:from-red-600 hover:to-red-700 transition-all duration-300 shadow-md hover:shadow-lg cursor-pointer flex items-center justify-center mt-9"
-            >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="h-5 w-5 mr-2"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1"
-                />
-              </svg>
-              CERRAR SESIÓN
-            </button>
+            <Logout logout={logout} />
           </div>
         </div>
       </aside>
